Fixed an infinite loop when cancelling the connection request on the authentication screen
When cancelling the auth screen, we would be caught in an infinite loop, because a redirect would add the provider back in the URL, + the HybridAuth storage was not properly cleared.

Note that Google does not seem to provide a way to cancel/deny the auth screen, which is not too cool.